Decoding Algorithm Complete Time Complexity: O(n) Space Complexity: O(n) [auxiliary space: O(n)]
I read that the decoding was quite difficult part, but I just implemented the inverse of the encoding method and it worked. The code took roughly another half hour to write but another half hour to debug. I used the array instead of the ArrayList this time, acting on the realization from the encoding part that the number of elements is known. Looking at other solutions to the problem, I see others using a single method returned by both the encoder and decoder. It makes sense since they both essentially follow the same process; I suppose that would be cleaner code and more in line with SOLID principles.